+++ schema/index.js
@@ -10,6 +10,7 @@ module.exports = {
     tiles: fs.readFileSync(path.join(__dirname, 'TABLE', 'tiles.sql'), 'utf8')
   },
   INDEX: {
-    metadata: fs.readFileSync(path.join(__dirname, 'INDEX', 'metadata.sql'), 'utf8')
+    metadata: fs.readFileSync(path.join(__dirname, 'INDEX', 'metadata.sql'), 'utf8'),
+    tiles: fs.readFileSync(path.join(__dirname, 'INDEX', 'tiles.sql'), 'utf8')
   }
 }